Release Nichidai2016 V2.0

+++ src/main/java/net/mybluemix/controller/HelloController.java
@@ -37,7 +37,7 @@ public class HelloController {
 
 	@Autowired
 	private SimpMessagingTemplate simpmessage;
-	
+
 	@RequestMapping("/test")
 	public VisualClassification greeting() {
 		VisualRecognition service = new VisualRecognition(VisualRecognition.VERSION_DATE_2016_05_20);
@@ -54,7 +54,7 @@ public class HelloController {
 
 	@RequestMapping(value = "/up", method = RequestMethod.POST)
 	public String post(@RequestParam MultipartFile multipartFile, WebRequest request) throws IOException {
-		
+
 		// ファイルが空の場合は異常終了
 		if (multipartFile.isEmpty()) {
 			// 異常終了時の処理
@@ -64,32 +64,36 @@ public class HelloController {
 			VisualRecognition service = new VisualRecognition(VisualRecognition.VERSION_DATE_2016_05_20);
 			service.setApiKey("ee08899658034e2f4ca599d2c2ba32da6eaa3435");
 			String suffix = getSuffix(multipartFile.getOriginalFilename());
-			String fName = DigestUtils.md5DigestAsHex((multipartFile.getOriginalFilename()+request.getSessionId()).getBytes()) + "." + suffix;
-			
+			String fName = DigestUtils.md5DigestAsHex(
+					(multipartFile.getOriginalFilename() + request.getSessionId()).getBytes()) + "." + suffix;
+
 			File imageF = convert(multipartFile, fName);
 			File out = new File(System.getProperty("user.dir") + "/temp/temp" + fName);
 			out.createNewFile();
 			BufferedImage image = ImageIO.read(imageF);
 			JPEGImageWriteParam param = new JPEGImageWriteParam(Locale.getDefault());
 			param.setCompressionMode(ImageWriteParam.MODE_EXPLICIT);
-			param.setCompressionQuality(0.8f);
+			param.setCompressionQuality(0.9f);
 
 			ImageWriter writer = ImageIO.getImageWritersByFormatName(suffix).next();
 			writer.setOutput(ImageIO.createImageOutputStream(out));
 			writer.write(null, new IIOImage(image, null, null), param);
 			writer.dispose();
+			if ((out.length() / 1024.0 / 1024.0) > 2.0) {
+				return "{\"out\":\"" + String.format("%.3f", (out.length() / 1024.0 / 1024.0)) + "MB" + "\"}";
+			}
 
-//			System.out.println("Classify an image");
+			// System.out.println("Classify an image");
 			ClassifyImagesOptions options = new ClassifyImagesOptions.Builder().images(out).build();
 			VisualClassification result = service.classify(options).execute();
 			VisualRecognitionOptions vop = new VisualRecognitionOptions.Builder().images(out).build();
 			DetectedFaces df = service.detectFaces(vop).execute();
-			
-			//tempファイル削除
+
+			// tempファイル削除
 			imageF.delete();
 			out.delete();
-//			System.out.println(result);
-			return "{\"result\":["+result.toString()+","+df.toString()+"]}";
+			// System.out.println(result);
+			return "{\"result\":[" + result.toString() + "," + df.toString() + "," + out.length() + "]}";
 		}
 
 	}
@@ -102,26 +106,27 @@ public class HelloController {
 		fos.close();
 		return convFile;
 	}
-	
+
 	public void makeTempDir() {
 		File mkdir = new File(System.getProperty("user.dir") + "/temp");
-		if(!mkdir.exists()) mkdir.mkdirs();
+		if (!mkdir.exists())
+			mkdir.mkdirs();
 	}
-	
+
 	public static String getSuffix(String fileName) {
-	    if (fileName == null)
-	        return null;
-	    int point = fileName.lastIndexOf(".");
-	    if (point != -1) {
-	        return fileName.substring(point + 1);
-	    }
-	    return fileName;
+		if (fileName == null)
+			return null;
+		int point = fileName.lastIndexOf(".");
+		if (point != -1) {
+			return fileName.substring(point + 1);
+		}
+		return fileName;
 	}
-	
-	@RequestMapping(value = "/dengon", method = RequestMethod.POST ,consumes = MediaType.APPLICATION_JSON_VALUE)
-	public String dengonController(@RequestBody Dengon dengon){
+
+	@RequestMapping(value = "/dengon", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public String dengonController(@RequestBody Dengon dengon) {
 		simpmessage.convertAndSend("/topic/dengon", dengon);
 		return "test";
 	}
-	
+
 }
